move utils into a module and split request_helpers and url_helpers
